requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==7.0.4
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pyngrok==7.5.1
pytesseract==0.3.13
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
